# src/app/api/hr/openapi.yaml
openapi: 3.0.0
info:
  title: HR & Asset Management API
  description: API documentation for the Human Resources & Asset Management module of the Hospital Management System
  version: 1.0.0
  contact:
    name: HMS Support
    email: support@hms.example.com
servers:
  - url: /api
    description: API base URL
...
paths:
...
  /hr/attendance/checkout:
    post:
      summary: Record checkout
      description: Records an attendance checkout
      tags:
        - Attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employeeId
              properties:
                employeeId:
                  type: string
                  description: Employee ID
                checkoutTime:
                  type: string
                  format: date-time
                  description: Checkout time (defaults to current time if not provided)
                notes:
                  type: string
                  description: Additional notes
      responses:
        '200':
          description: Checkout recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
        '400':
          description: Validation error or no active check-in found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /hr/payroll/periods:
    get:
      summary: List payroll periods
      description: Returns a paginated list of payroll periods
      tags:
        - Payroll
      parameters:
        - name: skip
          in: query
          description: Number of records to skip for pagination
          schema:
            type: integer
            default: 0
        - name: take
          in: query
          description: Number of records to return per page
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [DRAFT, PROCESSING, APPROVED, PAID]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods:
                    type: array
                    items:
                      $ref: '#/components/schemas/PayrollPeriod'
                  total:
                    type: integer
                  skip:
                    type: integer
                  take:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create payroll period
      description: Creates a new payroll period
      tags:
        - Payroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayrollPeriodInput'
      responses:
        '200':
          description: Payroll period created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollPeriod'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /hr/biomedical/{id}:
    get:
      summary: Get biomedical equipment
      description: Returns a specific biomedical equipment by ID
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiomedicalEquipment'
        '404':
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update biomedical equipment
      description: Updates an existing biomedical equipment
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiomedicalEquipmentUpdateInput'
      responses:
        '200':
          description: Biomedical equipment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiomedicalEquipment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete biomedical equipment
      description: Deletes a biomedical equipment
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      responses:
        '200':
          description: Biomedical equipment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hr/biomedical/{id}/calibration:
    get:
      summary: Get calibration records
      description: Returns calibration records for a specific biomedical equipment
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalibrationRecord'
        '404':
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Record calibration
      description: Records calibration for a specific biomedical equipment
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalibrationInput'
      responses:
        '200':
          description: Calibration recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrationRecord'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        joinDate:
          type: string
          format: date
        terminationDate:
          type: string
          format: date
        isActive:
          type: boolean
        department:
          $ref: '#/components/schemas/Department'
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        qualifications:
          type: array
          items:
            $ref: '#/components/schemas/Qualification'
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [CLINICAL, ADMINISTRATIVE, SUPPORT]
        description:
          type: string
        parentDepartmentId:
          type: string
        parentDepartment:
          $ref: '#/components/schemas/Department'
    Position:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        employeeId:
          type: string
        departmentId:
          type: string
        department:
          $ref: '#/components/schemas/Department'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isPrimary:
          type: boolean
...
    Qualification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [DEGREE, LICENSE, CERTIFICATION, TRAINING]
        name:
          type: string
        issuedBy:
          type: string
        issuedDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        attachmentUrl:
          type: string
        verificationStatus:
          type: string
          enum: [PENDING, VERIFIED, REJECTED]
...
    Attendance:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        employee:
          $ref: '#/components/schemas/Employee'
        date:
          type: string
          format: date
        checkinTime:
          type: string
          format: date-time
        checkoutTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [PRESENT, ABSENT, LATE, HALF_DAY]
        notes:
          type: string
        verificationMethod:
          type: string
          enum: [BIOMETRIC, MANUAL, SYSTEM]
...
    PayrollPeriod:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        paymentDate:
          type: string
          format: date
        status:
          type: string
          enum: [DRAFT, PROCESSING, APPROVED, PAID]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        paidBy:
          type: string
        paidAt:
          type: string
          format: date-time
    PayrollPeriodInput:
      type: object
      required:
        - name
        - startDate
        - endDate
        - paymentDate
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        paymentDate:
          type: string
          format: date
        status:
          type: string
          enum: [DRAFT]
          default: DRAFT
...
    Asset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        assetType:
          type: string
          enum: [EQUIPMENT, FURNITURE, IT, VEHICLE, OTHER]
        serialNumber:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          format: float
        warrantyExpiryDate:
          type: string
          format: date
        location:
          type: string
        departmentId:
          type: string
        department:
          $ref: '#/components/schemas/Department'
        assignedToId:
          type: string
        assignedTo:
          $ref: '#/components/schemas/Employee'
        status:
          type: string
          enum: [AVAILABLE, IN_USE, UNDER_MAINTENANCE, DISPOSED, LOST]
        lastMaintenanceDate:
          type: string
          format: date
        nextMaintenanceDate:
          type: string
          format: date
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        maintenanceRecords:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceRecord'
        assetHistory:
          type: array
          items:
            $ref: '#/components/schemas/AssetHistory'
...
    MaintenanceRecord:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
        maintenanceType:
          type: string
          enum: [PREVENTIVE, CORRECTIVE, CALIBRATION, INSPECTION]
        date:
          type: string
          format: date
        performedBy:
          type: string
        cost:
          type: number
          format: float
        description:
          type: string
        nextMaintenanceDate:
          type: string
          format: date
...
    AssetHistory:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
        type:
          type: string
          enum: [ACQUISITION, ASSIGNMENT, STATUS_CHANGE, MAINTENANCE, DISPOSAL]
        date:
          type: string
          format: date-time
        employeeId:
          type: string
        employee:
          $ref: '#/components/schemas/Employee'
        details:
          type: object
    BiomedicalEquipment:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
        asset:
          $ref: '#/components/schemas/Asset'
        equipmentType:
          type: string
          enum: [DIAGNOSTIC, THERAPEUTIC, MONITORING, LABORATORY, SURGICAL, LIFE_SUPPORT, OTHER]
        deviceIdentifier:
          type: string
        regulatoryClass:
          type: string
          enum: [CLASS_I, CLASS_II, CLASS_III]
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        lastCalibrationDate:
          type: string
          format: date
        nextCalibrationDate:
          type: string
          format: date
        calibrationFrequency:
          type: integer
        certifications:
          type: array
          items:
            type: string
        isReusable:
          type: boolean
        sterilizationRequired:
          type: boolean
        lastSterilizationDate:
          type: string
          format: date
        calibrationRecords:
          type: array
          items:
            $ref: '#/components/schemas/CalibrationRecord'
...
    BiomedicalEquipmentUpdateInput:
      type: object
      properties:
        name:
          type: string
        equipmentType:
          type: string
          enum: [DIAGNOSTIC, THERAPEUTIC, MONITORING, LABORATORY, SURGICAL, LIFE_SUPPORT, OTHER]
        serialNumber:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          format: float
        warrantyExpiryDate:
          type: string
          format: date
        location:
          type: string
        departmentId:
          type: string
        assignedToId:
          type: string
        status:
          type: string
          enum: [AVAILABLE, IN_USE, UNDER_MAINTENANCE, DISPOSED, LOST]
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        deviceIdentifier:
          type: string
        regulatoryClass:
          type: string
          enum: [CLASS_I, CLASS_II, CLASS_III]
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        lastCalibrationDate:
          type: string
          format: date
        nextCalibrationDate:
          type: string
          format: date
        calibrationFrequency:
          type: integer
        certifications:
          type: array
          items:
            type: string
        isReusable:
          type: boolean
        sterilizationRequired:
          type: boolean
        lastSterilizationDate:
          type: string
          format: date
    CalibrationRecord:
      type: object
      properties:
        id:
          type: string
        biomedicalEquipmentId:
          type: string
        date:
          type: string
          format: date
        performedBy:
          type: string
        result:
          type: string
          enum: [PASS, FAIL, ADJUSTED]
        notes:
          type: string
        attachments:
          type: array
          items:
            type: string
    CalibrationInput:
      type: object
      required:
        - date
        - result
      properties:
        date:
          type: string
          format: date
        performedBy:
          type: string
        result:
          type: string
          enum: [PASS, FAIL, ADJUSTED]
        notes:
          type: string
        nextCalibrationDate:
          type: string
          format: date
        attachments:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: object